add File sample

+++ js/gen/runScript_concat.js
@@ -1,4 +1,4 @@
-// Created at Wed Apr 30 2014 14:57:21 GMT+0900 (東京 (標準時))
+// Created at Fri May 02 2014 15:31:51 GMT+0900 (東京 (標準時))
 (function () {
 	var R={};
 	R.def=function (reqs,func,type) {
@@ -5133,13 +5133,17 @@ requirejs(["fs/ROMk","FS","Tonyu.Project","Shell","KeyEventChecker","ScriptTagFS
             h=$(window).height();
         	cv.attr({width: w, height: h});
         }
-        var curProjectDir=FS.get("/Tonyu/runscript/");
+	var locs=location.href.replace(/\?.*/,"").split(/\//);
+	var loc=locs.pop();
+	if (loc.length<0) locs="runscript";
+        var curProjectDir=FS.get("/Tonyu/"+loc+"/");
         if (curProjectDir.exists()) sh.rm(curProjectDir,{r:1});
         var fo=ScriptTagFS.toObj();
         for (var fn in fo) {
         	var f=curProjectDir.rel(fn);
            	f.text(fo[fn]);
         }
+	sh.cd(curProjectDir);
         var main="Main";
         var scrs=$("script");
         scrs.each(function (){
